Update ATLAS_OpenData_13-TeV__analysis_example-cpp_Hyy_channel.ipynb
Added a warning message for the reader into brackets, at line 198. This would help the reader to understand that if the running of the next line of code will take a few minutes this would be perfectly normal. First time I run that part of code I wasn't sure about that (becaude it took me more time than what I expected) so I refreshed the page and lost a few minutes for this.