Update HZZ_NN.ipynb
Added capital letters at the beginning of lines 305 and 324.
Added missing final dot at line 378.
Added a warning message for the reader into brackets, at line 503. This would help the reader to understand that if the running of the next line of code will take a few minutes this would be perfectly normal.