File caching in Higgs notebooks
For the two example notebooks to be used in the tutorial tomorrow (we
will likely only get through HZZ, but I wanted to get this into both
just in case), I'm adding here a (default True) option to download the
files to local storage and open them from there.

When remote-streaming the files using the https protocol, uproot seems
to generate a very large number of server requests, which can overwhelm
a webserver (particularly when several people are doing that at once).
There is a discussion here of how that could be fixed, but that
discussion hasn't yet converged:
https://github.com/scikit-hep/uproot5/discussions/1452

I'm preparing this PR so that we can merge it this evening, in
preparation for the tutorial tomorrow, if the uproot caching discussion
doesn't converge, so that we have something working out of the box
tomorrow.