Update to Systematics Notebook
This is a final draft of the systematics notebook. I have removed the section on Accuracy and Precision since it did not apply elsewhere in this notebook. I have a left a section on the Rules of Propagation of Uncertainties which show some formulas; I may remove this section since we don't use any of these formulas.

This update also includes a coding section, in which we look at calibration uncertainties in jets, scale factor uncertainties in muons, and modeling systematic uncertainties with jets. This last one may have an error, since one of the samples show very little counts compared to the other two.