Update to Acceptance/Efficiency Notebook
Made some minor changes to this notebook. Included a chart emoji in front of each histrogram title to help distinguish it from other sections. Revised some sentences and added a bit more sentences in some areas to improve flow and understandability. Images may still be acting weird, but it may be fixed once it is merged into the master branch.