Latest Update to Systematics and Uncertainty Notebook
This is the most recent version of the systematics notebook. I decided to remove the Accuracy and Precision section since it was unneeded in the remainder of the notebook. I may also remove the section that shows the formulas for propagation of uncertainties since that is also not used. I included examples on why uncertainties are needed in measurements, so feel free to let me know how well these examples are.

This update also includes a coding section, in which we look at calibration uncertainties in jets, scale factor uncertainties in muons, and modeling systematic uncertainties with jets. This last one may have an error, since one of the samples show very little counts compared to the other two.