Changing the logic of the masks and +documentation
While making the presentation for the workshop I thought that it's a bit odd to write masks that filter the contrary of what we are telling people we want, to then negate it when slicing. So I wrote the masks for what we want to keep. It should be easier to understand and explain.

I also added a point both the cut in the relative pT, that wasn't explained in the text. I wrote the reason I think this cut is done, but I might be wrong.