Adding NCB notebook
This is a first attempt at a notebook that shows the concept of
non-collision backgrounds. I'd like to expand it some day to show
both non-collision backgrounds and data quality issues, but I had
trouble coming up with an example of a data quality issue that is
sufficiently easy to look at with the E&O open data or that runs
fast enough with the research open data.